+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v1.4.20
+- Add to use chunks in tempstore (1500)
+- Added button to corrupt images
+- Fix unescaped pathname for ffprobe
+- Fix first image of album
+  
 ### v1.4.19
 - Fix wrong naming for videos
 - Fix naming scheme